refactor: move ImageAnalysisResultEntitySerializer
move ImageAnalysisResultEntitySerializer into Utilities folder
